pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Edward Castle <sir.edwardcastle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "1.11.1"
django-webpack-loader = "0.6.0"
djangorestframework-jwt = "1.11.0"
graphene = "2.1.8"
django-stubs = "0.10.1"
#djangorestframework-stubs = "0.4.2"
poetry = "^1.0.5"
cx-Oracle = "6.0"
graphene-django = "^2.13.0"
djangorestframework = "^3.9.1"
htmlmin = "^0.1.12"
django-htmlmin = "^0.11.0"
django-cors-headers = "^2.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
